# openapi.yaml
openapi: 3.0.3
info:
  title: DivvyDoo API
  description: API for expense splitting and group payment management
  version: 1.0.0
  contact:
    name: DivvyDoo Team
servers:
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Register a new user account
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user-lookup:
    get:
      tags:
        - Users
      summary: Look up user by email or phone
      description: Search for a user by their email address or phone number. Returns limited user info for privacy.
      operationId: lookupUser
      parameters:
        - name: q
          in: query
          required: true
          description: Email address or phone number to search for
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLookupResult'
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get user profile by user ID. Users can only access their own profile.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - cannot access other user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update user profile. Users can only update their own profile.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - cannot update other user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /settlements:
    post:
      tags:
        - Settlements
      summary: Create a settlement
      description: Create a settlement record to track a payment between users. The from_user_id must match the authenticated user.
      operationId: createSettlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettlementRequest'
      responses:
        '201':
          description: Settlement created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - can only create settlements from yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settlements/{id}:
    get:
      tags:
        - Settlements
      summary: Get settlement details
      description: Get settlement details by settlement ID. User must be involved in the settlement.
      operationId: getSettlement
      parameters:
        - name: id
          in: path
          required: true
          description: Settlement ID
          schema:
            type: string
      responses:
        '200':
          description: Settlement retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in the settlement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Settlement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settlements/{id}/complete:
    put:
      tags:
        - Settlements
      summary: Complete a settlement
      description: Mark a settlement as completed. User must be involved in the settlement.
      operationId: completeSettlement
      parameters:
        - name: id
          in: path
          required: true
          description: Settlement ID
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteSettlementRequest'
      responses:
        '200':
          description: Settlement completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in the settlement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Settlement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settlements/{id}/cancel:
    put:
      tags:
        - Settlements
      summary: Cancel a settlement
      description: Cancel a pending settlement. User must be involved in the settlement.
      operationId: cancelSettlement
      parameters:
        - name: id
          in: path
          required: true
          description: Settlement ID
          schema:
            type: string
      responses:
        '200':
          description: Settlement cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in the settlement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Settlement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settlements/pending:
    get:
      tags:
        - Settlements
      summary: Get pending settlements
      description: Get all pending settlements for the authenticated user.
      operationId: getPendingSettlements
      responses:
        '200':
          description: Pending settlements retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Settlement'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (minimum 8 characters)
          example: secretpassword123
        phone:
          type: string
          description: User's phone number
          example: "+1234567890"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        phone:
          type: string
          description: User's phone number
          example: "+1234567890"
...
    CreateSettlementRequest:
      type: object
      required:
        - from_user_id
        - to_user_id
        - amount
        - currency
        - method
      properties:
        from_user_id:
          type: string
          description: User ID of the payer (must match authenticated user)
          example: usr_abc123
        to_user_id:
          type: string
          description: User ID of the recipient
          example: usr_def456
        group_id:
          type: string
          description: Group ID (optional)
          example: grp_abc123
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Settlement amount
          example: 50.00
        currency:
          type: string
          description: Currency code
          example: USD
        method:
          type: string
          enum:
            - cash
            - bank_transfer
            - upi
            - paypal
            - venmo
            - other
          description: Payment method
          example: bank_transfer
        description:
          type: string
          description: Optional description
          example: Paying back for dinner
    CompleteSettlementRequest:
      type: object
      properties:
        transaction_id:
          type: string
          description: Optional external transaction ID
          example: txn_12345

    # Response Schemas
...
    UserLookupResult:
      type: object
      properties:
        user_id:
          type: string
          description: User ID
          example: usr_abc123
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
    User:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439011
        user_id:
          type: string
          description: User ID
          example: usr_abc123
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        phone:
          type: string
          description: User's phone number
          example: "+1234567890"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    Settlement:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439011
        settlement_id:
          type: string
          description: Settlement ID
          example: stl_abc123
        from_user_id:
          type: string
          description: User ID of the payer
          example: usr_abc123
        to_user_id:
          type: string
          description: User ID of the recipient
          example: usr_def456
        group_id:
          type: string
          description: Group ID (optional)
          example: grp_abc123
        amount:
          type: number
          format: double
          description: Settlement amount
          example: 50.00
        currency:
          type: string
          description: Currency code
          example: USD
        status:
          type: string
          enum:
            - pending
            - completed
            - failed
            - cancelled
          description: Settlement status
          example: pending
        method:
          type: string
          enum:
            - cash
            - bank_transfer
            - upi
            - paypal
            - venmo
            - other
          description: Payment method
          example: bank_transfer
        description:
          type: string
          description: Optional description
          example: Paying back for dinner
        transaction_id:
          type: string
          description: External transaction ID
          example: txn_12345
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        completed_at:
          type: string
          format: date-time
          description: Completion timestamp
        failed_at:
          type: string
          format: date-time
          description: Failure timestamp
        failure_reason:
          type: string
          description: Reason for failure
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid request body